Remove op register
