+++ brainpy/_src/math/__init__.py
@@ -47,7 +47,6 @@ from .others import *
 from . import random, linalg, fft
 
 # operators
-from .op_register import *
 from .pre_syn_post import *
 from . import surrogate, event, sparse, jitconn
