--- brainpy/_src/math/op_register/__init__.py
@@ -1,7 +1,0 @@
-from .numba_approach import (CustomOpByNumba,
-                             register_op_with_numba_xla,
-                             compile_cpu_signature_with_numba)
-from .base import XLACustomOp
-from .utils import register_general_batching
-from .base import XLACustomOp
-from .utils import register_general_batching
