+++ brainpy/_src/math/sparse/__init__.py
@@ -1,9 +1,7 @@
 # from ._coo_mv import *
-# from ._bsr_mv import *
 from .csr_mv import *
 from .csr_mm import *
 from .utils import *
-from .bsr_mm import *
 from .jax_prim import *
