--- brainpy/math/op_register.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-from brainpy._src.math.op_register import (
-  CustomOpByNumba,
-  compile_cpu_signature_with_numba,
-)
-
-from brainpy._src.math.op_register.base import XLACustomOp
-from brainpy._src.math.op_register.ad_support import defjvp
-
-
